requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
httpx
supabase
psycopg2-binary
